+++ Tests/SentryTests/SentryCrash/SentryCrashWrapperTests.swift
@@ -1,4 +1,5 @@
 @_spi(Private) @testable import Sentry
+@_spi(Private) import SentryTestUtils
 import XCTest
 
 final class SentryCrashWrapperTests: XCTestCase {
@@ -129,7 +130,7 @@ final class SentryCrashWrapperTests: XCTestCase {
             
             let runtimeContext = try XCTUnwrap(scope.contextDictionary["runtime"] as? [String: Any])
             XCTAssertEqual(runtimeContext["name"] as? String, "Mac Catalyst App")
-            XCTAssertEqual(runtimeContext["raw_description"] as? String, "raw_description")
+            XCTAssertEqual(runtimeContext["raw_description"] as? String, "mac-catalyst-app")
             #endif
     }
 }